requirements.txt
# Core dependencies
streamlit>=1.28.0
neo4j>=5.14.0
python-dotenv>=1.0.0
graphiti-core

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
pypandoc>=1.11  # Python wrapper for pandoc
Pillow>=9.0.0  # Image processing library (PIL)
PyMuPDF>=1.23.0  # PDF processing library (fitz)

# AI/ML (for custom implementation)
openai>=1.0.0
anthropic>=0.25.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-experimental>=0.3.4

# Data processing
numpy>=1.24.0
pydantic>=2.0.0

# Visualization
plotly>=5.17.0
pyvis>=0.3.2
networkx>=3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
chardet>=5.0.0
tiktoken>=0.5.0

# File processing utilities
python-magic>=0.4.27  # File type detection
chardet>=5.0.0       # Character encoding detection

# Optional KG libraries (uncomment if needed)
graphiti-core>=0.9.5        # Zep Graphiti - Advanced temporal KG (ENABLED)
# neo4j-graphrag>=0.3.0       # Neo4j's official GraphRAG  
# knowledge-graph-maker>=0.2.1 # Alternative KG library 
wand >=0.6.10  # Image processing for EMF/X-EMF conversion

pytz>=2025.1